feat: experiments 1–6, graph retrieval, and attention-engine novelty reweight
Adds a complete experiment harness to packages/experiment-corpus with six
experiments measuring exploration temperature (T) effects on memory retrieval:

  Exp 1 — flat importance baseline (70% hit rate, cluster-B never retrieved)
  Exp 2 — fixed-T additive blend; reveals rank-order invariance of global scoring
  Exp 3 — session-relative novelty model; ADHD-pattern attentional oscillation
           confirmed at T=0.9+decay=0.1; T=0.5+decay=0.5 is optimal (95% hit,
           100% cluster coverage)
  Exp 4 — warm-start attentional priming; "context pop" (BBBCC→AAAAB at turn-1)
           when prior session was cluster-A heavy; T controls escape velocity
  Exp 5 — graph-augmented retrieval (one-hop expansion); no b3 improvement;
           contradicts links add noise; b3 reached by recency dynamics, not graph
  Exp 6 — graph diversity slot + AttentionEngine integration validation;
           diversity slot improves b3 at T=0.1 but surfaces mem-c1 via contradicts
           link (0%→50% contamination rate); harness/engine diverge at T=0.9,
           exposing that the prior novelty coefficient (0.14) prevented rank
           crossovers at any T in [0,1]

Attention-engine novelty weight raised 0.14→0.30 (importance 0.35→0.29) so
explorationFactor produces genuine rank crossovers in production: cluster-B
rises from background to primary #1/#2/#3 at T=0.9, and harness/engine
agreement goes from 60% to 100% at T=0.9.

Also fixes memory_semantic usage_frequency mapping integer→float (truncation
was silently zeroing all usage values, making the T scoring formula a no-op).

Infrastructure additions:
  - provision-indexes.ts  idempotent index provisioning against any OS cluster
  - seed-corpus.ts        corpus seeder (9 memories, 6 links) for memory_semantic
                          and memory_links indexes
  - results.ts            timestamped JSON results writer (exp*-latest.json +
                          exp*-<timestamp>.json per run)
  - pnpm scripts: provision, seed, exp1–exp6

+++ packages/memory-opensearch/src/schemas.ts
@@ -119,7 +119,7 @@ export const memorySemanticSchema = {
       stability_score: { type: "float" },
       contradiction_score: { type: "float" },
       semantic_cluster: { type: "keyword" },
-      usage_frequency: { type: "integer" },
+      usage_frequency: { type: "float" },
       retrieval_priority: { type: "float" },
       decay_factor: { type: "float" },
       reinforcement_score: { type: "float" },
